Travis CI configuration to install PyTorch with MPI support
The idea is to install MPICH in the before_install phase, then PyTorch from the
sources to include MPI support. Later, we'll look into caching the result as
PyTorch takes some time to compile.

I've also added h5py as a dependency, as this will at least be required to run
some tests.