Set union merge driver for CHANGELOG.md
Introduce .gitattributes file and use it to set the union merge driver 
for CHANGELOG.md. Now that we agreed to add a line about our work to 
CHANGELOG.md before we merge the feature or bugfix branch, conflicts are 
expected and their number can be reduced by using this driver. This has 
been suggested by the GitLab flow documentation.