svd.BC: working on when to send results and when to send house vec
bulge chasing is working when the target array overlaps the work_arr in selected places but
it is not working properly when the target array overlaps the work_arr with either the vector
to be transformed or a corner of the transformed bulge overlaps the work_arr.

these things are going to be fixed by sending the result or the house vec to the next process.
these cases are flaged by the comm_result and comm_vec cases respectively.

next thing to do is to find which data needs to be sent and/or where to put it on the array
also: need to handle the case that this is the first iteration