Bugfix: nprocs > shape[1]
Having more processes than chunks in split 1 did not work. Rather than 
checking whether we are the last (overall) rank, we check whether we 
have the last chunk of data and don't write anything if we have no data. 
Last chunk is relevant to distinguish newline or separator addition at 
the end of our buffer.