Migrate to NVidia PyTorch base image
NVidia images come with support for HPC systems desirable for our uses.
They work a little differently internally and required some changes.

The tzdata configuration configures the CET/CEST timezone, which seems
to be required when installing additional packages.

There is an issue with pip caches in the image, which led to the final
cache purge to fail in the PyPI release based build. This is fixed
through a final invocation of true.