Fix tzdata handling and merging multiple actions
The copy of tzdata was originally intended to be used from the root
of the repository. The documentation states building docker images
from the docker directory. This had to be aligned.

Moreover, each RUN line in a Dockerfile represents the building of
another intermediate image. This has to be taken into account when
writing the files. Commands such as "cd" will not work on separate
lines.